[fe][feat] 펀딩 메인 api 분리 완료
