+++ FE/src/components/Funding/api/FundingAPI.jsx
@@ -113,29 +113,6 @@ async function createFunding(formData, token) {
   return await response.json();
 }
 
-//펀딩 피드 api
-async function getFundingFeed(token, setData) {
-  try {
-    const response = await axios.get(
-      `${import.meta.env.VITE_BASE_URL}/api/fundings/feed`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-        params: {
-          page: 0,
-          size: 100,
-          sort: "", // sort가 필요하다면 'columnName,asc' 또는 'columnName,desc' 형식의 값을 설정하세요.
-        },
-      },
-    );
-    console.log("펀딩 피드 가져오기 : ", response.data.data.data);
-    setData(response.data.data.data);
-  } catch (error) {
-    console.error("펀딩 피드를 불러올 수 없습니다.", error);
-  }
-}
-
 //펀딩 삭제 api
 
 async function deleteFunding(token, fundingId, navigate) {
@@ -167,6 +144,5 @@ export {
   fetchFriendFunding,
   fetchMyFundings,
   fetchDetailFunding,
-  getFundingFeed,
   deleteFunding,
 };
